Conflicts:
	matsim/src/main/java/org/matsim/core/mobsim/qsim/qnetsimengine/QueueWithBuffer.java

+++ playgrounds/boescpa/src/main/java/playground/boescpa/av/staticDemand/AutonomousVehicle.java
@@ -124,7 +124,7 @@ public class AutonomousVehicle {
         }
     }
 
-    public void incServiceDistance(long serviceDistance) {
+    public void incServiceDistance(double serviceDistance) {
         if (serviceDistance >= 0) {
             totalServiceDistance += serviceDistance;
         } else {
